requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.4.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
cachelib==0.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
chatterbot-corpus==1.2.0
clarifai==2.6.2
click==8.0.1
colorama==0.4.4
configparser==3.8.1
cryptography==3.3.1
cssselect==1.1.0
decorator==4.4.2
Deprecated==1.2.12
discord.py==1.6.0
disrank==0.0.2
EasyProcess==0.3
emoji==1.2.0
future==0.18.2
google-trans-new==1.1.9
googleapis-common-protos==1.52.0
googletrans==4.0.0rc1
grpcio==1.35.0
h11==0.9.0
h2==3.2.0
howdoi==2.0.18
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
huepy==1.2.1
hyperframe==5.2.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
iniconfig==1.1.1
instabot==0.117.0
instant-api-client==0.1.2
instapy==0.6.13
isort==5.7.0
joblib==1.0.1
jsonschema==2.6.0
keep==2.10.1
language-tool-python==2.5.3
lazy-object-proxy==1.4.3
lxml==4.6.3
mccabe==0.6.1
MeaningCloud-python==2.0.0
mock==4.0.3
MouseInfo==0.1.3
moviepy==1.0.3
multidict==5.1.0
mysql-connector==2.2.9
mysql-connector-python==8.0.28
newsapi-python==0.2.6
nltk==3.6.2
nmap==0.0.1
numpy==1.20.1
packaging==20.9
pefile==2019.4.18
Pillow==8.1.0
pluggy==0.13.1
plyer==2.0.0
proglog==0.1.9
protobuf==3.14.0
prsaw==0.4.0
py==1.10.0
PyAutoGUI==0.9.52
pycparser==2.20
pygame==2.0.1
PyGetWindow==0.0.9
PyGithub==1.55
Pygments==2.10.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
PyJWT==2.1.0
pylint==2.6.0
PyMsgBox==1.0.9
PyNaCl==1.4.0
pynput==1.7.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
pyquery==1.4.3
PyRect==0.1.4
PyScreeze==0.1.27
PySocks==1.7.1
pyspellchecker==0.6.2
pytest==6.2.2
python-telegram-bot==13.2
PyTweening==1.0.3
pytz==2021.1
pywhatkit==4.7
pywin32-ctypes==0.2.0
PyYAML==3.13
randomstuff.py==2.0.0
regex==2020.11.13
requests==2.25.1
requests-toolbelt==0.9.1
responses==0.12.1
rfc3986==1.5.0
schedule==1.0.0
selenium==3.141.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.1
terminaltables==3.1.0
textblob==0.15.3
toml==0.10.2
tornado==6.1
tqdm==4.56.0

tzlocal==2.1



wrapt==1.12.1

